// src/common/interfaces/userWord.tsx
import { Difficulty } from "../enums/difficulty";

export interface IUserWord {
  id: string;
  group: number;
  page: number;
  word: string;
  image: string;
  audio: string;
  audioMeaning: string;
  audioExample: string;
  textMeaning: string;
  textExample: string;
  transcription: string;
  wordTranslate: string;
  textMeaningTranslate: string;
  textExampleTranslate: string;
  difficulty: Difficulty;
  goals: number;
  isUserWord: boolean;
}